create user attributes in user account
